+++ statistics/common.c
@@ -44,7 +44,7 @@ double amath_stdev(double *data, size_t n_elements) {
   return sqrt(square_sigma / n_elements);
 }
 
-unsigned long long amath_factorial(unsigned int x) {
+unsigned long long amath_factorial(unsigned long long x) {
   if (x == 0) return 1;
   return amath_factorial(x - 1) * x;
 }
